requirements.txt
asgiref==3.6.0
boto3==1.26.52
botocore==1.29.52
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-bootstrap-datepicker-plus==5.0.3
django-storages==1.13.2
django-webpack-loader==1.8.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
node==1.2
odict==1.9.0
Pillow==9.4.0
plumber==1.7
pydantic==1.10.4
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==4.6.1
